mckinsey-ppt-generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.21
pydantic
pydantic-settings
python-dotenv==1.0.0
redis==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
celery==5.3.4
loguru==0.7.2
httpx==0.25.2
tenacity==8.2.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
bcrypt==4.1.1
Pillow
fonttools==4.53.1
anthropic==0.39.0
openai==1.6.1
tiktoken==0.5.2
aiosqlite
python-docx==1.1.0        # Word 문서 파싱
PyPDF2==3.0.1             # PDF 문서 파싱
markdown==3.5.1           # Markdown 파싱
psutil==5.9.6             # 프로세스 및 시스템 유틸리티
matplotlib==3.8.2         # 차트 생성
requests==2.32.3
pdfplumber==0.11.0        # PDF 텍스트 추출 향상
streamlit==1.36.0         # Streamlit UI
plotly==5.24.1            # 품질 검토 시각화 (Streamlit)
